+++ src/bigint/index.ts
@@ -1,4 +1,3 @@
-export * from './is-bigint'
 export * from './max-bigint'
 export * from './min-bigint'
 export * from './negative-bigint'