+++ src/type/is-null/is-null.predicate.ts
@@ -1,0 +1,7 @@
+/**
+ * @category Predicates
+ * @param value The value to validate.
+ */
+export function isNull(value: unknown): value is null {
+    return value === null
+}